pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "bloodwork-ai"
version = "0.1.0"
description = "Multi-modal bloodwork AI platform for image and tabular data analysis"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "bloodwork_ai", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.1.0"
torchvision = "^0.16.0"
ultralytics = "^8.0.0"
xgboost = "^2.0.0"
lightgbm = "^4.0.0"
scikit-learn = "^1.3.0"
optuna = "^3.4.0"
mlflow = "^2.8.0"
dvc = "^3.0.0"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pandas = "^2.1.0"
numpy = "^1.24.0"
albumentations = "^1.3.0"
shap = "^0.43.0"
grad-cam = "^1.4.0"
pycocotools = "^2.0.7"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
plotly = "^5.17.0"
jupyter = "^1.0.0"
jupyterlab = "^4.0.0"
prometheus-client = "^0.19.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
tqdm = "^4.66.0"
pillow = "^10.0.0"
opencv-python = "^4.8.0"
scipy = "^1.11.0"
imbalanced-learn = "^0.11.0"
calibration = "^0.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^23.0.0"
isort = "^5.12.0"
ruff = "^0.1.0"
mypy = "^1.7.0"
pre-commit = "^3.5.0"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["bloodwork_ai"]

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src/bloodwork_ai",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow tests",
]
